WIP: Extract methods from DataLayoutBuilder to utils
