+++ modules/apps/data-engine/data-engine-taglib/.npmbundlerrc
@@ -8,7 +8,7 @@
 				"/": ">=4.0.0"
 			},
 			"dynamic-data-mapping-form-renderer": {
-				"/": ">=4.0.0"
+				"/": ">=6.0.0"
 			}
 		}
 	},
