+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/components/field-sets/actions/useSaveFieldSet.es.js
@@ -104,7 +104,7 @@ export default ({
 									value: dataDefinitionFields.map(({name}) =>
 										dataLayoutBuilder.getDDMFormField(
 											childrenState.dataDefinition,
-											name
+											name // TODO: Remove unused param
 										)
 									),
 								},
