+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/handlers/fieldSetAddedHandler.es.js
@@ -47,7 +47,10 @@ const handleFieldSetAdded = (props, state, event) => {
 
 	let fieldSetField = createFieldSet(
 		props,
-		{skipFieldNameGeneration: false, useFieldName: useFieldName[props.defaultLanguageId]},
+		{
+			skipFieldNameGeneration: false,
+			useFieldName: useFieldName?.[props.defaultLanguageId],
+		},
 		nestedFields
 	);
